# src/entrypoints/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Advent of Code - OpenAPI 3.0",
    "description": "This is a simple backend API that solves the puzzle of the Advent Of code challenge",
    "contact": {
      "email": "https://github.com/minirulo"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "Advent of Code",
    "url": "https://adventofcode.com/"
  },
  "servers": [
    {
      "url": "http://localhost:8008"
    }
  ],
  "tags": [
    {
      "name": "elf_supply",
      "description": "Check how many calories do the elves transport",
      "externalDocs": {
        "description": "Puzzle",
        "url": "https://adventofcode.com/2022/day/1"
      }
    },
    {
      "name": "strategy",
      "description": "Get the result of a perfect Rock Paper Scissors strategy",
      "externalDocs": {
        "description": "Puzzle",
        "url": "https://adventofcode.com/2022/day/2"
      }
    },
    {
      "name": "elf_duty",
      "description": "Check how many duties overlap when the elves pair with each other",
      "externalDocs": {
        "description": "Puzzle",
        "url": "https://adventofcode.com/2022/day/4"
      }
    },
    {
      "name": "warehouse_stack",
      "description": "Check the last crates in the stack after the operatior follows the transactions",
      "externalDocs": {
        "description": "Puzzle",
        "url": "https://adventofcode.com/2022/day/5"
      }
    },
    {
      "name": "device",
      "description": "Find the position of the unique chain of characters within a message",
      "externalDocs": {
        "description": "Puzzle",
        "url": "https://adventofcode.com/2022/day/6"
      }
    }
  ],
  "paths": {
    "/api/v1/supplies/{elves}": {
      "post": {
        "tags": [
          "elf_supply"
        ],
        "parameters": [
          {
            "name": "elves",
            "in": "path",
            "description": "Top number of elves to get the calories from.",
            "required": false,
            "default": 1,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/strategy/{winning}": {
      "post": {
        "tags": [
          "strategy"
        ],
        "parameters": [
          {
            "name": "winning",
            "in": "path",
            "description": "Select wether to follow your instinct [false] or the winning strategy [true]",
            "required": false,
            "default": "false",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/elves/{overlap}": {
      "post": {
        "tags": [
          "elf_duty"
        ],
        "parameters": [
          {
            "name": "overlap",
            "in": "path",
            "description": "Whether to check for fully overlapped duties [true] or not [anything else]",
            "required": false,
            "default": "true",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stacks/{crateMover}": {
      "post": {
        "tags": [
          "warehouse_stack"
        ],
        "parameters": [
          {
            "name": "crateMover",
            "in": "path",
            "description": "Model of the CrateMover to use [ 9000 or 9001 ]",
            "required": false,
            "default": 9000,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/device/{length}": {
      "post": {
        "tags": [
          "device"
        ],
        "parameters": [
          {
            "name": "length",
            "in": "path",
            "description": "Length of the unique packet / message to look for in the message",
            "required": false,
            "default": 4,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "result": { }
        }
      }
    }
  }
}